apollo_starknet_client,apollo_rpc: delete writer module
Move DeprecatedContractClass into apollo_rpc's broadcasted_transaction module,
update api/mod.rs and tests to use it locally, then delete the entire writer
module from apollo_starknet_client.
