--- crates/apollo_starknet_client/src/writer/objects/mod.rs
@@ -1,4 +1,0 @@
-pub mod response;
-#[cfg(any(feature = "testing", test))]
-pub mod test_utils;
-pub mod transaction;
